Fix filtering webhook by labels
